Rename 'bitstar' folder to 'informedtrees' and add aitstar to it
