+++ demos/OptimalPlanning.cpp
@@ -43,7 +43,7 @@
 #include "ompl/util/Console.h"
 
 // The supported optimal planners, in alphabetical order
-#include <ompl/geometric/planners/bitstar/BITstar.h>
+#include <ompl/geometric/planners/informedtrees/BITstar.h>
 #include <ompl/geometric/planners/cforest/CForest.h>
 #include <ompl/geometric/planners/fmt/FMT.h>
 #include <ompl/geometric/planners/fmt/BFMT.h>
