+++ py-bindings/headers_geometric.txt
@@ -28,7 +28,7 @@ src/ompl/geometric/planners/rrt/LBTRRT.h
 src/ompl/geometric/planners/rrt/LazyLBTRRT.h
 src/ompl/geometric/planners/rrt/InformedRRTstar.h
 src/ompl/geometric/planners/rrt/SORRTstar.h
-src/ompl/geometric/planners/bitstar/BITstar.h
+src/ompl/geometric/planners/informedtrees/BITstar.h
 src/ompl/geometric/planners/sbl/SBL.h
 src/ompl/geometric/planners/stride/STRIDE.h
 src/ompl/geometric/planners/fmt/FMT.h
