+++ src/ompl/geometric/planners/AnytimePathShortening.cpp
@@ -41,7 +41,7 @@
 #include "ompl/geometric/planners/fmt/BFMT.h"
 #include "ompl/geometric/planners/est/BiEST.h"
 #include "ompl/geometric/planners/rrt/BiTRRT.h"
-#include "ompl/geometric/planners/bitstar/BITstar.h"
+#include "ompl/geometric/planners/informedtrees/BITstar.h"
 #include "ompl/geometric/planners/kpiece/BKPIECE1.h"
 #include "ompl/geometric/planners/est/EST.h"
 #include "ompl/geometric/planners/fmt/FMT.h"
