+++ src/ompl/geometric/planners/informedtrees/AITstar.h
@@ -34,8 +34,8 @@
 
 // Authors: Marlin Strub
 
-#ifndef OMPL_GEOMETRIC_PLANNERS_AITSTAR_AITSTAR_
-#define OMPL_GEOMETRIC_PLANNERS_AITSTAR_AITSTAR_
+#ifndef OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_AITSTAR_
+#define OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_AITSTAR_
 
 #include <algorithm>
 #include <memory>
@@ -43,9 +43,9 @@
 #include "ompl/base/Planner.h"
 #include "ompl/datastructures/BinaryHeap.h"
 #include "ompl/geometric/PathGeometric.h"
-#include "ompl/geometric/planners/aitstar/datastructures/Edge.h"
-#include "ompl/geometric/planners/aitstar/datastructures/ImplicitGraph.h"
-#include "ompl/geometric/planners/aitstar/datastructures/Vertex.h"
+#include "ompl/geometric/planners/informedtrees/aitstar/Edge.h"
+#include "ompl/geometric/planners/informedtrees/aitstar/ImplicitGraph.h"
+#include "ompl/geometric/planners/informedtrees/aitstar/Vertex.h"
 
 namespace ompl
 {
@@ -292,4 +292,4 @@ namespace ompl
     }  // namespace geometric
 }  // namespace ompl
 
-#endif  // OMPL_GEOMETRIC_PLANNERS_AITSTAR_AITSTAR
+#endif  // OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_AITSTAR
