+++ src/ompl/geometric/planners/informedtrees/bitstar/HelperFunctions.h
@@ -34,8 +34,8 @@
 
 /* Authors: Jonathan Gammell */
 
-#ifndef OMPL_GEOMETRIC_PLANNERS_BITSTAR_COMMONHELPERFUNCTIONS_
-#define OMPL_GEOMETRIC_PLANNERS_BITSTAR_COMMONHELPERFUNCTIONS_
+#ifndef OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_BITSTAR_HELPERFUNCTIONS_
+#define OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_BITSTAR_HELPERFUNCTIONS_
 
 // std::pair
 #include <utility>
@@ -92,4 +92,4 @@ namespace
     };
 }
 ////////////////////////////////
-#endif  // OMPL_GEOMETRIC_PLANNERS_BITSTAR_COMMONHELPERFUNCTIONS_
+#endif  // OMPL_GEOMETRIC_PLANNERS_INFORMEDTREES_BITSTAR_HELPERFUNCTIONS_
